Clean up xdg_config_dir and move some stuff to utils
