+++ Makefile
@@ -1,4 +1,4 @@
-SRC = main.c arg.c xdg.c
+SRC = main.c arg.c xdg.c utils.c
 CC ?= clang
 
 all: cnetstat
